feat(json-utils): share json format code with package
