+++ apps/vscode/src/extension.ts
@@ -1,5 +1,6 @@
 import * as vscode from "vscode";
 import * as fs from "fs/promises";
+import { getJsonContainer } from "@repo/json-utils";
 
 export function activate(context: vscode.ExtensionContext) {
   console.log('Congratulations, your extension "formatter-lab" is now active!');
@@ -165,63 +166,4 @@ function getWebViewContent(formattedJson: string): string {
         </body>
         </html>
     `;
-}
-
-function highlightJSON(value: any, depth = 0): string {
-  const indentClass = `depth-${depth}`;
-
-  if (Array.isArray(value)) {
-    const items = value
-      .map(
-        (item, index) =>
-          `<div class="json-itens ${indentClass}">
-            <span class="index">${index}:</span> ${highlightJSON(
-              item,
-              depth + 1
-            )}
-          </div>`
-      )
-      .join("");
-
-    return `
-      <span class="bracket">[</span>
-      <div class="${indentClass} json-content">${items}</div>
-      <span class="bracket">]</span>`;
-  }
-
-  if (value && typeof value === "object") {
-    const entries = Object.entries(value)
-      .map(
-        ([key, item]) =>
-          `<div class="json-itens ${indentClass}">
-            <span class="key">"${key}"</span>: ${highlightJSON(item, depth + 1)}
-          </div>`
-      )
-      .join("");
-
-    return `
-      <span class="bracket">{</span>
-      <div class="${indentClass} json-content">${entries}</div>
-      <span class="bracket">}</span>`;
-  }
-
-  if (value === null) {
-    return `<span class="value">null</span>`;
-  }
-
-  if (typeof value === "number") {
-    return `<span class="value">${value}</span>`;
-  }
-
-  if (typeof value === "string") {
-    return `<span class="value">"${value}"</span>`;
-  }
-
-  return `<span class="value">${value}</span>`;
-}
-
-function getJsonContainer(json: any) {
-  return `<section class="json-container"><div class="json">${highlightJSON(
-    json
-  )}</div></section>`;
 }
