+++ apps/web/src/scripts/json-tree-view.ts
@@ -1,58 +1,7 @@
+import { getJsonContainer } from "@repo/json-utils";
+
 const filePicker = document.getElementById("file-picker");
 const content = document.getElementById("content");
-
-function highlightJSON(value: any, depth = 0): string {
-  const indentClass = `depth-${depth}`;
-
-  if (Array.isArray(value)) {
-    const items = value
-      .map(
-        (item, index) =>
-          `<div class="json-itens ${indentClass}">
-            <span class="index">${index}:</span> ${highlightJSON(
-              item,
-              depth + 1
-            )}
-          </div>`
-      )
-      .join("");
-
-    return `
-      <span class="bracket">[</span>
-      <div class="${indentClass} json-content">${items}</div>
-      <span class="bracket">]</span>`;
-  }
-
-  if (value && typeof value === "object") {
-    const entries = Object.entries(value)
-      .map(
-        ([key, item]) =>
-          `<div class="json-itens ${indentClass}">
-            <span class="key">"${key}"</span>: ${highlightJSON(item, depth + 1)}
-          </div>`
-      )
-      .join("");
-
-    return `
-      <span class="bracket">{</span>
-      <div class="${indentClass} json-content">${entries}</div>
-      <span class="bracket">}</span>`;
-  }
-
-  if (value === null) {
-    return `<span class="value">null</span>`;
-  }
-
-  if (typeof value === "number") {
-    return `<span class="value">${value}</span>`;
-  }
-
-  if (typeof value === "string") {
-    return `<span class="value">"${value}"</span>`;
-  }
-
-  return `<span class="value">${value}</span>`;
-}
 
 function appendCollapseEvent() {
   const collapseBrackets = document.querySelectorAll(".bracket");
@@ -97,9 +46,7 @@ filePicker!.addEventListener("change", async (e) => {
 
   if (!jsonString) return;
 
-  content!.innerHTML = `<section class="json-container"><div class="json">${highlightJSON(
-    JSON.parse(jsonString)
-  )}</div></section>`;
+  content!.innerHTML = getJsonContainer(JSON.parse(jsonString));
 
   appendCollapseEvent();
 });
